requirements.txt
Flask==3.0.3
google-cloud-storage==2.16.0
gunicorn==22.0.0